Move utils.py -> plex_trakt_sync/util.py
